efficient frontier using knapsack optimization
i became concerned with performance, and so i also
implemented an approximation in the packer, but having
done so, it seems maybe not necessary.

anyway, the pareto curves look MUCH better now, and
i think they're pretty close to optimal.

note that none of this applies for the more-general
case of multiple cost constraints, but it should work,
with a little modification, for correlated values.

also, unit tests.  yay!