extract accessors from mapper code
now there are two layers of accessors:

* domain access.  accessor functions for the backbone
model and collection objects we already use, so the
actual field-name access is limited to that scope only

* dimension access.  for any particular visualization, we
choose domain dimensions (e.g. "Stage"); i made wrappers
for the domain models that do this mapping.

so now there are three layers:

1. raw domain object to domain accessors
2. domain accessors to chart dimensions
3. chart dimensions to chart dataset

the 'chart dataset' has a shape that is more-or-less
the same as the DOM we end up with.

all this is only true for the bingo chart, so far.

the motivation is to remove client-specific stuff from
the later layers.